Refactor Program.cs, moved logic to new Runner.cs
