+++ .gitignore
@@ -15,3 +15,4 @@ SBOM/
 .nuget/
 .opencode-server/
 TestResults/
+/CycloneDX/Properties/launchSettings.json
